pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-format-normalizer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "lxml>=5.2.0",
    "python-docx>=1.1.2",
    "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
]

[project.scripts]
paper-format-normalizer = "paper_format_normalizer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
